+++ x/emissions/types/events_test.go
@@ -48,7 +48,7 @@ func TestEmitNewInfererScoresSetEventWithScores(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventScoresSet", event.Type)
+	require.Equal(t, "emissions.v7.EventScoresSet", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -107,7 +107,7 @@ func TestEmitNewForecasterScoresSetEventWithScores(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventScoresSet", event.Type)
+	require.Equal(t, "emissions.v7.EventScoresSet", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -166,7 +166,7 @@ func TestEmitNewReputerScoresSetEventWithScores(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventScoresSet", event.Type)
+	require.Equal(t, "emissions.v7.EventScoresSet", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -225,7 +225,7 @@ func TestEmitNewInfererRewardsSettledEventWithRewards(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventRewardsSettled", event.Type)
+	require.Equal(t, "emissions.v7.EventRewardsSettled", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -284,7 +284,7 @@ func TestEmitNewForecasterRewardsSettledEventWithRewards(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventRewardsSettled", event.Type)
+	require.Equal(t, "emissions.v7.EventRewardsSettled", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -343,7 +343,7 @@ func TestEmitNewReputerAndDelegatorRewardsSettledEventWithRewards(t *testing.T) 
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventRewardsSettled", event.Type)
+	require.Equal(t, "emissions.v7.EventRewardsSettled", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -404,7 +404,7 @@ func TestEmitNewNetworkLossSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventNetworkLossSet", event.Type)
+	require.Equal(t, "emissions.v7.EventNetworkLossSet", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 3)
@@ -437,7 +437,7 @@ func TestEmitNewForecastTaskSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventForecastTaskScoreSet", event.Type)
+	require.Equal(t, "emissions.v7.EventForecastTaskScoreSet", event.Type)
 
 	require.Contains(t, event.Attributes[0].Key, "score")
 	require.Contains(t, event.Attributes[0].Value, "10")
@@ -460,9 +460,9 @@ func TestNewLastCommitSetEvent(t *testing.T) {
 	events := ctx.EventManager().Events()
 	require.Len(t, events, 3)
 
-	require.Equal(t, "emissions.v5.EventWorkerLastCommitSet", events[0].Type)
-	require.Equal(t, "emissions.v5.EventWorkerLastCommitSet", events[1].Type)
-	require.Equal(t, "emissions.v5.EventReputerLastCommitSet", events[2].Type)
+	require.Equal(t, "emissions.v7.EventWorkerLastCommitSet", events[0].Type)
+	require.Equal(t, "emissions.v7.EventWorkerLastCommitSet", events[1].Type)
+	require.Equal(t, "emissions.v7.EventReputerLastCommitSet", events[2].Type)
 
 	require.Contains(t, events[0].Attributes[0].Key, "block_height")
 	require.Contains(t, events[0].Attributes[1].Key, "nonce")
@@ -499,7 +499,7 @@ func TestEmitNewTopicRewardsSetEvent(t *testing.T) {
 	events := ctx.EventManager().Events()
 	require.Len(t, events, 1)
 
-	require.Equal(t, "emissions.v5.EventTopicRewardsSet", events[0].Type)
+	require.Equal(t, "emissions.v7.EventTopicRewardsSet", events[0].Type)
 	require.Contains(t, events[0].Attributes[0].Key, "rewards")
 	require.Contains(t, events[0].Attributes[0].Value, `["0","10","20","30","40"]`)
 	require.Contains(t, events[0].Attributes[1].Key, "topic_ids")
@@ -538,7 +538,7 @@ func TestEmitNewEMAScoresSetEventWithScores(t *testing.T) {
 	require.Len(t, events, 3)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventEMAScoresSet", event.Type)
+	require.Equal(t, "emissions.v7.EventEMAScoresSet", event.Type)
 
 	require.Contains(t, events[0].Attributes[0].Key, "actor_type")
 	require.Contains(t, events[0].Attributes[0].Value, "\"ACTOR_TYPE_INFERER_UNSPECIFIED\"")
@@ -579,7 +579,7 @@ func TestEmitNewListeningCoefficientsSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventListeningCoefficientsSet", event.Type)
+	require.Equal(t, "emissions.v7.EventListeningCoefficientsSet", event.Type)
 
 	attributes := event.Attributes
 	require.Len(t, attributes, 5)
@@ -646,7 +646,7 @@ func TestEmitNewInfererNetworkRegretSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventInfererNetworkRegretSet", event.Type)
+	require.Equal(t, "emissions.v7.EventInfererNetworkRegretSet", event.Type)
 
 	val, exists := event.GetAttribute(AttributeKeyTopicId)
 	require.True(t, exists)
@@ -703,7 +703,7 @@ func TestEmitNewForecasterNetworkRegretSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventForecasterNetworkRegretSet", event.Type)
+	require.Equal(t, "emissions.v7.EventForecasterNetworkRegretSet", event.Type)
 
 	val, exists := event.GetAttribute(AttributeKeyTopicId)
 	require.True(t, exists)
@@ -759,7 +759,7 @@ func TestEmitNewNaiveInfererNetworkRegretSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventNaiveInfererNetworkRegretSet", event.Type)
+	require.Equal(t, "emissions.v7.EventNaiveInfererNetworkRegretSet", event.Type)
 
 	val, exists := event.GetAttribute(AttributeKeyTopicId)
 	require.True(t, exists)
@@ -814,7 +814,7 @@ func TestEmitNewTopicInitialRegretSetEvent(t *testing.T) {
 	require.Len(t, events, 1)
 
 	event := events[0]
-	require.Equal(t, "emissions.v5.EventTopicInitialRegretSet", event.Type)
+	require.Equal(t, "emissions.v7.EventTopicInitialRegretSet", event.Type)
 
 	val, exists := event.GetAttribute(AttributeKeyTopicId)
 	require.True(t, exists)